+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### 2.3.5 (Unreleased)
 
+- [#550](https://github.com/influxdata/clockface/pull/550): Introduce `ButtonGroup` component
+- [#550](https://github.com/influxdata/clockface/pull/550): Make `ConfirmationButton` appear in active state while popover is visible
 - [#543](https://github.com/influxdata/clockface/pull/543): Add Funnel, Bucket, and Function icons to icon font
 - [#541](https://github.com/influxdata/clockface/pull/541): Improve `testID` usability in `TreeNav` component