+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### 2.3.5 (Unreleased)
 
+- [#551](https://github.com/influxdata/clockface/pull/551): Introduce `TimeInput` component
 - [#550](https://github.com/influxdata/clockface/pull/550): Introduce `ButtonGroup` component
 - [#550](https://github.com/influxdata/clockface/pull/550): Make `ConfirmationButton` appear in active state while popover is visible
 - [#543](https://github.com/influxdata/clockface/pull/543): Add Funnel, Bucket, and Function icons to icon font